Fix Go WASM error with node Crypto module
On the latest versions of Node we started seeing errors about
`globalThis` not having the `crypto` module.

I'm not sure _why_ this started happening. But it appears we can fill in
the globalThis by requiring the `crypto` package and filling it in
manually.

I also changed the imports so that